+++ gui_logic.py
@@ -8,12 +8,14 @@ from kivy.clock import Clock
 from kivy.uix.screenmanager import ScreenManager, Screen
 from kivy.uix.label import Label
 from kivy.uix.button import Button
+from kivy.metrics import dp
 
-from common_utils import LOCAL_CALENDAR, get_month, get_month_name
+from common_utils import LOCAL_CALENDAR, get_month, get_month_name, get_week_number, get_week_days
 
 TODAY = datetime.today()
 
 class CalendarScreens(ScreenManager):
+    """ Switches between different views - Up/Down (y axis) movement """
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self._touch_start_y = 0
@@ -26,14 +28,18 @@ class CalendarScreens(ScreenManager):
         dy = touch.y - self._touch_start_y
         if abs(dy) > 50:
             if dy > 0:
-                self.go_next()
+                new_screen = self.next()
+                self.transition.direction = 'up'
             else:
-                self.go_previous()
+                new_screen = self.previous()
+                self.transition.direction = 'down'
+            self.current = new_screen
         return super().on_touch_up(touch)
 
 class YearScreens(ScreenManager):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+        # self.current = 'current_year'
         self._touch_start_x = 0
 
     def on_touch_down(self, touch):
@@ -43,15 +49,19 @@ class YearScreens(ScreenManager):
     def on_touch_up(self, touch):
         dx = touch.x - self._touch_start_x
         if abs(dx) > 50:
-            if dx > 0:
-                self.go_next()
+            if dx < 0:
+                new_screen = self.next()
+                self.transition.direction = 'left'
             else:
-                self.go_previous()
+                new_screen = self.previous()
+                self.transition.direction = 'right'
+            self.current = new_screen
         return super().on_touch_up(touch)
 
 class MonthScreens(ScreenManager):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+        # self.current = 'current_month'
         self._touch_start_x = 0
 
     def on_touch_down(self, touch):
@@ -61,15 +71,19 @@ class MonthScreens(ScreenManager):
     def on_touch_up(self, touch):
         dx = touch.x - self._touch_start_x
         if abs(dx) > 50:
-            if dx > 0:
-                self.go_next()
+            if dx < 0:
+                new_screen = self.next()
+                self.transition.direction = 'left'
             else:
-                self.go_previous()
+                new_screen = self.previous()
+                self.transition.direction = 'right'
+            self.current = new_screen
         return super().on_touch_up(touch)
 
 class WeekScreens(ScreenManager):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+        # self.current = 'current_week'
         self._touch_start_x = 0
 
     def on_touch_down(self, touch):
@@ -79,40 +93,79 @@ class WeekScreens(ScreenManager):
     def on_touch_up(self, touch):
         dx = touch.x - self._touch_start_x
         if abs(dx) > 50:
-            if dx > 0:
-                self.go_next()
+            if dx < 0:
+                new_screen = self.next()
+                self.transition.direction = 'left'
             else:
-                self.go_previous()
+                new_screen = self.previous()
+                self.transition.direction = 'right'
+            self.current = new_screen
         return super().on_touch_up(touch)
 
 class CurrentYearScreen(Screen):
     """ Year View - show months & buttons for each month """
-    def on_enter(self, *args) -> None:
+    def on_pre_enter(self, *args) -> None:
         Clock.schedule_once(self.update_label, 0)
 
     def update_label(self, dt) -> None:
-        self.ids.current_year_label.text = 2025 # placeholder
         self.ids.current_year_grid.clear_widgets()
+
+        year: int = 2025 # placeholder
+        self.ids.current_year_label.text = f'{year}'
 
         for month in range(1, 13):
             month_name = get_month_name(month)
             btn = Button(text=month_name)
             btn.bind(on_release=lambda instance, m=month: self.on_month_selected(m))
-            self.ids.year_grid.add_widget(btn)
+            self.ids.current_year_grid.add_widget(btn)
 
     def on_month_selected(self, month) -> None:
         """ Load MonthView """
         print('call month')
 
 class PrevYearScreen(Screen):
-    pass
+    def on_pre_enter(self, *args) -> None:
+        Clock.schedule_once(self.update_label, 0)
+
+    def update_label(self, dt) -> None:
+        self.ids.prev_year_grid.clear_widgets()
+
+        year: int = 2024
+        self.ids.prev_year_label.text = f'{year}'
+
+        for month in range(1, 13):
+            month_name = get_month_name(month)
+            btn = Button(text=month_name)
+            btn.bind(on_release=lambda instance, m=month: self.on_month_selected(m))
+            self.ids.prev_year_grid.add_widget(btn)
+
+    def on_month_selected(self, month) -> None:
+        """ Load MonthView """
+        print('call month')
 
 class NextYearScreen(Screen):
-    pass
+    def on_pre_enter(self, *args) -> None:
+        Clock.schedule_once(self.update_label, 0)
+
+    def update_label(self, dt) -> None:
+        self.ids.next_year_grid.clear_widgets()
+
+        year: int = 2026
+        self.ids.next_year_label.text = f'{year}'
+
+        for month in range(1, 13):
+            month_name = get_month_name(month)
+            btn = Button(text=month_name)
+            btn.bind(on_release=lambda instance, m=month: self.on_month_selected(m))
+            self.ids.next_year_grid.add_widget(btn)
+
+    def on_month_selected(self, month) -> None:
+        """ Load MonthView """
+        print('call month')
 
 class CurrentMonthScreen(Screen):
     """ Populate MonthView """
-    def on_enter(self, *args) -> None:
+    def on_pre_enter(self, *args) -> None:
         """ Delay load until widget tree is complete """
         Clock.schedule_once(self.update_label, 0)
 
@@ -151,9 +204,7 @@ class CurrentMonthScreen(Screen):
         # day_view.open()
 
 class PrevMonthScreen(Screen):
-    """ Populate MonthView """
-    def on_enter(self, *args) -> None:
-        """ Delay load until widget tree is complete """
+    def on_pre_enter(self, *args) -> None:
         Clock.schedule_once(self.update_label, 0)
 
     def update_label(self, dt) -> None:
@@ -161,9 +212,14 @@ class PrevMonthScreen(Screen):
 
         # placeholder values
         year: int = 2025
-        month: int = 4
+        month: int = 3
 
-        self.ids.prev_month_label.text = 'PreviousMonthView'
+        self.ids.prev_month_label.text = f"{LOCAL_CALENDAR.formatmonthname(
+            theyear=year,
+            themonth=month,
+            width=0,
+            withyear=True
+        )}"
 
         for day in range(7):
             self.ids.prev_month_grid.add_widget(Label(
@@ -186,19 +242,23 @@ class PrevMonthScreen(Screen):
         # day_view.open()
 
 class NextMonthScreen(Screen):
-    """ Populate MonthView """
-    def on_enter(self, *args) -> None:
-        """ Delay load until widget tree is complete """
+    def on_pre_enter(self, *args) -> None:
         Clock.schedule_once(self.update_label, 0)
 
     def update_label(self, dt) -> None:
+        self.ids.next_month_label.clear_widgets()
         self.ids.next_month_grid.clear_widgets()
 
         # placeholder values
         year: int = 2025
-        month: int = 4
+        month: int = 5
 
-        self.ids.next_month_label.text = 'PreviousMonthView'
+        self.ids.next_month_label.text = f"{LOCAL_CALENDAR.formatmonthname(
+            theyear=year,
+            themonth=month,
+            width=0,
+            withyear=True
+        )}"
 
         for day in range(7):
             self.ids.next_month_grid.add_widget(Label(
@@ -221,12 +281,72 @@ class NextMonthScreen(Screen):
         # day_view.open()
 
 class CurrentWeekScreen(Screen):
-    pass
+    def on_pre_enter(self, *args) -> None:
+        Clock.schedule_once(self.update_label, 0)
+
+    def update_label(self, dt) -> None:
+        self.ids.current_week_days.clear_widgets()
+
+        year: int = 2025
+        current_day = datetime.today()
+        week: str = get_week_number(current_day=current_day)
+        week_list: list = get_week_days(current_day=current_day)
+
+        self.ids.current_week_label.text = f'{year}\nWeek {week}'
+
+        for day in range(7):
+            self.ids.current_week_days.add_widget(Label(
+                text=LOCAL_CALENDAR.formatweekday(day=(day + 6) % 7, width=3)
+            ))
+        for date in week_list:
+            self.ids.current_week_days.add_widget(Label(
+                text=f'{date}', font_size=dp(13)
+            ))
 
 class PrevWeekScreen(Screen):
-    pass
+    def on_pre_enter(self, *args) -> None:
+        Clock.schedule_once(self.update_label, 0)
+
+    def update_label(self, dt) -> None:
+        self.ids.prev_week_days.clear_widgets()
+
+        year: int = 2025
+        current_day = datetime.today()
+        week: str = get_week_number(current_day=current_day)
+        week_list: list = get_week_days(current_day=current_day)
+
+        self.ids.prev_week_label.text = f'{year}\nWeek {week}'
+
+        for day in range(7):
+            self.ids.prev_week_days.add_widget(Label(
+                text=LOCAL_CALENDAR.formatweekday(day=(day + 6) % 7, width=3)
+            ))
+        for date in week_list:
+            self.ids.prev_week_days.add_widget(Label(
+                text=f'{date}', font_size=dp(13)
+            ))
 
 class NextWeekScreen(Screen):
-    pass
+    def on_pre_enter(self, *args) -> None:
+        Clock.schedule_once(self.update_label, 0)
+
+    def update_label(self, dt) -> None:
+        self.ids.next_week_days.clear_widgets()
+
+        year: int = 2025
+        current_day = datetime.today()
+        week: str = get_week_number(current_day=current_day)
+        week_list: list = get_week_days(current_day=current_day)
+
+        self.ids.next_week_label.text = f'{year}\nWeek {week}'
+
+        for day in range(7):
+            self.ids.next_week_days.add_widget(Label(
+                text=LOCAL_CALENDAR.formatweekday(day=(day + 6) % 7, width=3)
+            ))
+        for date in week_list:
+            self.ids.next_week_days.add_widget(Label(
+                text=f'{date}', font_size=dp(13)
+            ))
 
 GUI = Builder.load_file('main.kv')